Remove parallel code in format, increasing performance by about 3x
This is most likely due to a poor implementation, and may return in the future. There was also added a target field to the format context struct, which might be used in the future as a output target, in the hopes of decreasing the high number of string allocations currently found in the format code. The target field is currently unused.